add mario get in pipe event

+++ 05-SceneManager/GameObject.h
@@ -36,7 +36,7 @@ public:
 	void SetSpeed(float vx, float vy) { this->vx = vx, this->vy = vy; }
 	void GetPosition(float &x, float &y) { x = this->x; y = this->y; }
 	void GetSpeed(float &vx, float &vy) { vx = this->vx; vy = this->vy; }
-	void SetNX(int nx) { this->nx = nx; }
+	void SetNX(float nx) { this->nx = nx; }
 	int GetState() { return this->state; }
 	virtual void Delete() { isDeleted = true;  	}
 	bool IsDeleted() { return isDeleted; }
